Fix Game to reflect the game state
I'm trying to follow the recomendations provided by
the How to Design Programs - 2nd Edition (HTDP2e) book.
The recipe is meant to be used to design programs in a functional style,
but I want to explore to which extent the HTDP2e recommendations
can improve my object-oriented design skills.

Scrabble is an interactive program (a world program in the HTDP2e
vocabulary.) According to HTDP2e, the first step in the design of
such programs consists in defining a data type to hold the state
of the program. The structure of the data will then shape the
structure of the program. For us, `Game` is that structure.

See http://www.ccs.neu.edu/home/matthias/HtDP2e and
particularly the end of the 6.1 section for the design recipe.